K8s Lib Injection: Extract scenario
